eslint issues fixed, yay!!!
