+++ package.json
@@ -36,6 +36,10 @@
 		"enzyme-adapter-react-16": "^1.12.1",
 		"enzyme-to-json": "^3.3.5",
 		"eslint": "^5.16.0",
+		"eslint-config-airbnb": "^17.1.0",
+		"eslint-plugin-import": "^2.17.2",
+		"eslint-plugin-jsx-a11y": "^6.2.1",
+		"eslint-plugin-react": "^7.13.0",
 		"flow-bin": "^0.98.1"
 	},
 	"jest": {
