+++ src/App.jsx
@@ -1,0 +1,32 @@
+/* eslint-disable no-tabs */
+// @flow
+
+import React from 'react';
+import logo from './logo.svg';
+import './App.css';
+import Ipv4Calc from './app/ipv4.react';
+
+function App() {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+        <p>
+					Edit
+          <code>src/App.js</code>
+					and save to reload.
+
+        </p>
+        <a className="App-link" href="https://reactjs.org" target="_blank" rel="noopener noreferrer">
+					Learn React
+
+        </a>
+      </header>
+      <content>
+        <Ipv4Calc />
+      </content>
+    </div>
+  );
+}
+
+export default App;
